refactor: move handle selection to provider
